# Conflicts:
#	app/Providers/AppServiceProvider.php

+++ database/migrations/2020_11_09_075331_create_business_plans_table.php
@@ -14,7 +14,7 @@ class CreateBusinessPlansTable extends Migration
     public function up()
     {
         Schema::create('business_plan', function (Blueprint $table) {
-            $table->integer('id');
+            $table->bigIncrements('id');
             $table->string('name');
             $table->string('description');
             $table->integer('maximum_storage_file');
